Fix issue #1113 - Move Exit.Cause to top Level
