+++ core/jvm/src/test/scala-2.12/zio/StacktracesSpec.scala
@@ -45,8 +45,8 @@ class StacktracesSpec(implicit ee: org.specs2.concurrent.ExecutionEnv)
   "single effectTotal for-comprehension" >> singleEffectTotalForComp
   "single suspendWith for-comprehension" >> singleSuspendWithForComp
 
-  private def show(trace: ZTrace): Unit        = if (debug) println(trace.prettyPrint)
-  private def show(cause: Exit.Cause[_]): Unit = if (debug) println(cause.prettyPrint)
+  private def show(trace: ZTrace): Unit   = if (debug) println(trace.prettyPrint)
+  private def show(cause: Cause[_]): Unit = if (debug) println(cause.prettyPrint)
 
   private def mentionsMethod(method: String, trace: ZTraceElement): Boolean =
     trace match {
@@ -79,7 +79,7 @@ class StacktracesSpec(implicit ee: org.specs2.concurrent.ExecutionEnv)
   private def mentionedMethod(method: String): Matcher[List[ZTraceElement]] = mentionMethod(method)
 
   private implicit final class CauseMust[R >: Environment](io: ZIO[R, _, _]) {
-    def causeMust(check: Exit.Cause[_] => Result): Result =
+    def causeMust(check: Cause[_] => Result): Result =
       unsafeRunSync(io).fold[Result](
         cause => {
           show(cause)
