+++ core/shared/src/main/scala/zio/FiberFailure.scala
@@ -16,8 +16,6 @@
 
 package zio
 
-import zio.Exit.Cause
-
 /**
  * Represents a failure in a fiber. This could be caused by some non-
  * recoverable error, such as a defect or system error, by some typed error,
